Project 1/app/requirements.txt
# Alpaca SDK (trading, data, news)
alpaca-py>=0.21.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Machine Learning - HMM
hmmlearn>=0.3.0
scikit-learn>=1.3.0

# NLP - Sentiment Analysis
transformers>=4.35.0
torch>=2.1.0

# Technical Analysis helpers (minimal - we implement manually)
# (no external TA library needed)

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Visualization
matplotlib>=3.8.0

# RSS fallback
feedparser>=6.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# httpx>=0.25.0

